Refactor: strategy 도메인 추가
