+++ src/main/java/step5/strategy/MovingStrategy.java
@@ -1,4 +1,4 @@
-package step5;
+package step5.strategy;
 
 public interface MovingStrategy {
