+++ src/main/java/step5/strategy/RandomDrivingStrategy.java
@@ -1,4 +1,7 @@
-package step5;
+package step5.strategy;
+
+import step5.strategy.MovingStrategy;
+import step5.util.NumberGenerator;
 
 public class RandomDrivingStrategy implements MovingStrategy {
